requirements.txt
# ---------- Core ----------
numpy>=1.23
pandas>=2.0
pillow>=10.0
tqdm>=4.66
openpyxl>=3.1
pyyaml>=6.0

# ---------- Vision / CV ----------
opencv-python>=4.10.0
matplotlib>=3.8
seaborn>=0.13
scipy>=1.11            # for Hungarian matching in grid_snap

# ---------- Deep Learning ----------
# Install torch that matches your CUDA; see note below.
torch                   # pinned by environment (see install note)
transformers==4.40.*    # matches the logs you’re running with
accelerate>=0.28.0      # used by Trainer path; safe to keep even if "min" script
peft>=0.10.0,<0.12      # LoRA adapters compatible with transformers 4.40
safetensors>=0.4.3
huggingface_hub>=0.23.0

# ---------- Metrics (optional but recommended) ----------
evaluate>=0.4.1         # for CER metric
# (evaluate pulls in datasets/etc.)

# ---------- Detection ----------
ultralytics==8.*        # YOLOv8/11

# ---------- Server / Tools ----------
Flask>=3.0,<4.0
